refactor into a nicer format
